Add files via upload
# Overview
The dataset consists of  instances of iris flowers, with each instance having 4 numerical attributes and one nominal attribute representing the species of the flower. The primary goal is to classify the iris species based on these attributes.
# Dataset Information
Relation: ZETECH
Attributes:
sepal_length (real): Length of the sepal in centimeters.
sepal_width (real): Width of the sepal in centimeters.
petal_length (real): Length of the petal in centimeters.
petal_width (real): Width of the petal in centimeters.
species (nominal): The species of the iris flower. Possible values: {Iris-setosa, Iris-versicolor, Iris-virginica}
# Tools and Technologies
WEKA: Used for data analysis, preprocessing, and machine learning model building.
ARFF Format: The dataset is provided in the Attribute-Relation File Format (ARFF) which is native to WEKA.